[HUDI-995] Move HoodieTestDataGenerator to hudi-common
